created init methods which can be used in any activity, refactored code, configured dashboard header, fixed dashboard
Added method initViews(), initSharedPreferences() and initNavigation() to Dashboard. initViews() purpose is to put all the view initialization there.
InitSharedPreferences is to initialize the shared preferences in order to be able to access the logged user. initNavigation is for better code quality. All
initialization for the navigation is now coded in that method.